pyproject.toml
[project]
name = "ask-volk"
version = "0.0.1"
license = { file = "LICENSE" }
authors = [{ name = "Yann Billeter" }]
description = "A tool to predict the outcome of public votes."
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "pyro-ppl==1.9.0",
    "requests==2.31",
    "dvc==3.49",
    "dvc_s3==3.1.0",
    "openpyxl==3.1.2",
    "hydra-core==1.3.2",
    "pandas[parquet]==2.2.2",
    "numpy==1.26.4",
    "alive-progress==3.1.5",
    "xgboost==2.1.0",
]

[project.optional-dependencies]
dev = ["ruff==0.4.1", "pytest==8.1.1"]

[project.scripts]
ask-volk-downloader = "ask_volk.downloader:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
packages = ["ask_volk"]
include = ["ask_volk/config/**/*.yaml"]
exclude = ["tests", "resources", "data"]

[tool.pytest.ini_options]
addopts = ["--import-mode=importlib"]

[tool.ruff]
line-length = 100
select = ["E", "F", "D"]          # "D" includes pydocstyle rules
ignore = ["D200", "D212", "D415"]

[tool.ruff.pydocstyle]
convention = "google"

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true
warn_return_any = true
warn_unused_configs = true
